+++ Sources/SuperwallKit/Misc/Extensions/NotificationName+Superwall.swift
@@ -1,0 +1,12 @@
+//
+//  Notifications.swift
+//  
+//
+//  Created by Jake Mor on 4/27/23.
+//
+
+//import Foundation
+//
+//extension Notification.Name {
+//    public static let superwallIsInitialized = Notification.Name("com.superwall.shared.isInitialized")
+//}